Implement mode switch for the console;
Up until now, Vedeu has used a 'cooked' mode terminal, meaning that characters as typed are output to the terminal until return is pressed. Pressing return signified the end of the 'command' and this input was then sent onwards throughout the program.

This led to a situation whereby simple applications which accept just keypress input could not exist- each keypress had to be accompanied with the respective return key to trigger execution.

The new default is for key presses to be interpreted directly- i.e. a 'raw' mode terminal. Pressing 'escape' now switches to 'cooked' mode, allowing more elaborate commands. The caveat being that in order to go back to raw mode, one has to type escape followed by return.