Remove Command stuff;
Now that there are events, this functionality is no longer required.

This simplifies the code and the whole process. It has removed the need to both the Output module and Process class; instead we simply trigger a ':refresh' event. We no longer need to enqueue the keypress, so that has been removed from Input.

On receipt of the :refresh event, each interface outputs directly to the terminal.