Remove Repository stuff;
Since we are only storing interfaces, there is no need for all this infrastructure. Replacing it is a simple class: Persistence. Also, as a side effect of the repository removal, we no longer need all the builder code- we can keep a single class which provides the convenience DSL for the creation of interfaces.

- Moved builder/builder to support/builder
- Remove repositories/ directory.