Add event throttling;
A delay can be specified when defining an event to ensure that the event is not triggered any more often than the delay. For example; a delay of 0.5 would mean that an event can only be triggered once every 500ms. This is handy when it is known before-hand that an event may be triggered very frequently. The request to trigger is effectively discarded at this time.