Handle centred interface geometry better;
We cannot use top and left in viewport_height and viewport_width respectively, as their values depend on the current viewport size which will change if the terminal is resized.

x, y, height and width are now pretty much reserved for holding the values the client wanted to use, but should not be used to do any rendering.