Allow changes in geometry;
The geometry of an interface may have been changed in a particular view, this change allows the interface to be re-drawn with these new attributes. The original attributes, which were defined when the application started are preserved.

Something to consider here is that maybe a front and back buffer is insufficient to map an interfaces attributes between views. If an interface is changing shape for some reason, how do we go back to previous or even the original settings?