Minor template related improvements;
During this small spike, I've started on the idea of adding view helpers to be processed in the ERb templates. This has given rise to two main concerns: 1) When the template should be processed; at the moment this is immediately, but maybe this should be deferred until last moment since 2) future helpers may add colours and styles, resulting in escape sequences which is incompatible with the way that Vedeu renders views).

- Replace `template.erb` with individual interface ERb templates.
- Add #time_now helper method to experiment with helpers in ERb templates.
- Use #split_lines instead of text.split.