Add RefreshCursor class;
This class will handle the refreshing of the cursor for the interface in focus or the terminal at large. If the cursor's offset positions are outside the interface's bounds, then it will trigger a refresh of the interface to ensure it is redrawn. This has a side-effect of being a bit of a performance win, since for large interfaces, the cursor will be inside the interface, (with no offset) for longer, meaning we don’t have to redraw the entire interface when the cursor moves.).