Move :tick and :tock event triggers;
This allows the timing of the client application from the point of keypress or command invocation (if in cooked mode) to the point a new application cycle begins.

Moving around the cursor example app, the time differences reported are ~5ms, however once we force the interface to update by moving out of the initial boundary, the speed drops to ~35ms. For 60 fps, we need this to be 15ms. Interesting.