Minor optimisation for clearing the interface;
If the interface is the same size as the terminal, we can gain a significant performance boost by using the terminal clear screen escape code, (\e[2J). If the interface is smaller than the terminal, at the moment, I see no need to create Char objects for each 'clear' character (space) that we want to write to the terminal.

Potential downsides to this approach is when the output is not destined for the terminal, but instead will be rendered as HTML for example.