Allow user to change project #base_path
ApplicationView#template and the loading code in Vedeu::Bootstrap
implicitly depend on files being located relative to cwd. This commit
adds a base_path configuration variable which *defaults* to cwd
(`File.expand_path('.')`) and which can be set in client code.

This probably isn't a good ultimate default because this won't work for
most common application scenarios. I punted on implementing what seems
like the most reasonable default, automatically setting the base path
relative to the location of the 'client' code. This is actually harder
that it might seem at first. This is how Rails sets its root config
value:

https://github.com/rails/rails/blob/3e36db4406beea32772b1db1e9a16cc1e8aea14c/railties/lib/rails/application.rb#L364

The approach used there seems a little intrusive so I've deferred the
decision for later.