Handle general text input better;
Previously, it was not clear how to use the fake terminal mode and the
Vedeu::Editor collection of classes to create a general text input in a
client application.

This commit, (relating to issue #278) aims to both address the documentation
and implementation of custom editable areas of the terminal screen.

- Add more documentation to Input::Capture relating to :fake mode.
- Include some work towards addressing issues with Output::Text and correctly
  wrapping content.
- Remove 'test/support/examples/editor_app.rb' as 'examples/dsl_editor.rb'
  (shortly available) will make this example redundant.