Examples may fail because ./tmp is not accessible;
When cloning the repository, the ./tmp directory for the project is not created
because it is part of the .gitignore file. Since this was a local convenience
for me, the log directory in the examples has now been moved to use the system
'/tmp' directory which should be available to both MacOSX and Linux users.

Fixes #283.