Trigger events instead of calling methods here;
The benefit is that we get visibility in the log when an event is fired,
whereas method calls are silent. This gives an explanation to the multiple
updates for Vedeu::Cursors::Cursor objects in the log file- they were being
hidden before output and shown on completion- this was to prevent a flickering
cursor issue.