Add more documentation.
Mainly, this is trying to mark as many classes or modules as @api private;
meaning their usage is supposed to be by Vedeu internally. This is not
strict in any sense since some classes contain methods triggered by events and
if using any of the `Vedeu.*` methods in your library, they will probably rely
on the internals of these classes.

This is an attempt to reduce the documentation surface area- documentation will
still be added/updating in these classes, but as a developer of a client
application using Vedeu, it is not expected that you know the internals of
these classes- their information provided to be helpful rather than as a guide.