Log when events are triggered even with no action;
Triggering an event in Vedeu would normally have resulted in an entry in the log
file if the event had an action and produced a result. This has now been updated
to log even with no action, so you can see if there are events being triggered,
to which nothing is listening to. An example, for the time being is the
:_mouse_event_ which is not bound by Vedeu. A click will trigger this and a
message: "No action for: ':_mouse_event_'" will appear in the log file if
configured.