Trigger :key event when keypressed;
When a key is pressed in the :raw or :fake terminal modes,
trigger the :key event also so that listeners can 'do stuff' too.

When a key is pressed in the :cooked terminal mode, then the
:command event is triggered as in cooked mode, enter/return has to be
pressed to send the input to the terminal.

By binding to :key or :command, you can receive the last key or
collection of keys (in Vedeu, termed a 'command').

Work relating to #306.