Fix weird cursor bug;
For some reason the cursor was not being rendered in the view for some
interfaces because no name was being passed to the Cursors#by_name method.
The method signature in Repository states that if name is not provided, use
Vedeu.focus. However, if it is provided, but it is nil, then we would not
know which cursor to move. Silly Gav.