Don't create an @attributes ivar;
Previously, when using the Repositories::Defaults module, an @attributes
instance variable was created for the including class, this is unnecessary
and also has the annoying side-effect during development of duplicating
the attributes when inspecting the including class. If the attributes are
needed by consuming classes, then I would prefer to have an explicit
method for those.