Providing a width is no longer mandatory;
Within the view DSL, when building aligned lines, it was previously necessary
to provide a width when specifying an alignment of :centre (:center), or :right.

This is no longer mandatory, since the model name is given to Output::Text, and
therefore the geometry of the visible area can be determined.

The testing example at test/support/examples/material_colours_app.rb has been
updated to reflect this also.