Remove #name from DSL classes;
To build classes using the DSL, either a name must be given or it is inherited
as the class is built as part of another class being built in the DSL. Therefore,
there is no need to have the #name methods in the DSL to define the model name.

This simplifies the DSL only by a fraction.

- Also, add and update tests in this area.

Work relates to #342. (Found and remedied during investigation).