Remove dependencies 'pry' and 'ruby-prof';
Previously, pry and ruby-prof were dependencies of Vedeu to provide debugging
and profiling respectively. Since they are not needed for core operations of
Vedeu, they have been removed. The code that once relied on them has been
modified to check for their existence and raise an exception if they are not
present if these Vedeu API methods were called. The affected API methods are:
'Vedeu.debug' and 'Vedeu.profile'.