Fix ENV['TERM'] for tmux;
tmux by default sets the $TERM environment variable to 'screen' which Vedeu
does not recognise in order to set the colour mode. This results in a very
limited palette, which isn't brilliant. A workaround within tmux is to:

set -g default-terminal "screen-256color"

This sets the $TERM value for tmux to 'screen-256color', which with this
commit, Vedeu now recognises and treats the same as 'term-256color' which
provides a 256 colour palette.

Vedeu has also been updated to recognise 'screen' (in addition to 'xterm')
which means the 16 colour palette of the terminal should be used.

Fixes #343.