Implement 'escape' renderer;
This renderer converts escape sequences to string-symbols which look like that
which you would find building an abstract syntax tree parser. This is easier on
the eye and helps with debugging- and has the nice side effect of being able to
be 'diffed' on the command line without codging the terminal up.

'Codging' is Lincolnshire slang for 'ruining' or 'messing up' or 'corrupting'.