Set cursor y/x offsets to 1 rather than 0;
When set to 0, the initial move_right or move_down action will need to be
triggered twice to cause the cursor to appear to move in these directions.

This commit also adds more validation to movement of the cursor to prevent
unnecessary work when the cursor does not need to change position.